+++ app/lib/requests/createBody.js
@@ -3,8 +3,8 @@ const VError = require('verror');
 const constants = require('../constants');
 const log = require('../logger');
 
-function forGPSearch(query) {
-  const search = query.split(' ').map(s => `${s}*`).join(' ');
+function forGPSearch(locals) {
+  const search = locals.query.split(' ').map(s => `${s}*`).join(' ');
   return {
     count: true,
     filter: 'OrganisationTypeID eq \'GPB\'',
@@ -13,27 +13,28 @@ function forGPSearch(query) {
     highlightPreTag: '<span class="highlight">',
     search,
     searchFields: 'OrganisationName,Address1,Address2,Address3,Postcode',
-    select: 'OrganisationName,Address1,Address2,Address3,City,County,Postcode,CCG',
+    select: 'OrganisationName,Address1,Address2,Address3,City,County,Postcode,CCG,Longitude,Latitude',
     top: 25,
   };
 }
 
-function forIAPTSearch(query) {
+function forIAPTSearch(locals) {
   return {
-    filter: `ServiceCodesProvided/any(c:search.in(c, '${constants.IAPTServiceCode}')) and OrganisationTypeID ne 'TRU' and RelatedIAPTCCGs/any(c: c eq '${query}')`,
+    filter: `ServiceCodesProvided/any(c:search.in(c, '${constants.IAPTServiceCode}')) and OrganisationTypeID ne 'TRU' and RelatedIAPTCCGs/any(c: c eq '${locals.query}')`,
+    orderby: `geo.distance(Geocode, geography'Point(${locals.lon} ${locals.lat})')`,
   };
 }
 
-function createBody(type, query) {
+function createBody(type, locals) {
   switch (type) {
     case constants.types.GP: {
-      return forGPSearch(query);
+      return forGPSearch(locals);
     }
     case constants.types.IAPT: {
-      return forIAPTSearch(query);
+      return forIAPTSearch(locals);
     }
     default: {
-      log.error(`Unable to create body for uknown type: ${type} with query: ${query}`);
+      log.error(`Unable to create body for uknown type: ${type} with query: ${locals.query}`);
       throw new VError(`Unable to create body for unknown type: ${type}`);
     }
   }