+++ test/unit/lib/requests/createBody.js
@@ -9,7 +9,7 @@ describe('createBody', () => {
   describe('unknown type', () => {
     it('should throw VError', () => {
       const unknownType = 'unknown';
-      expect(() => createBody(unknownType, null)).to.throw(VError, `Unable to create body for unknown type: ${unknownType}`);
+      expect(() => createBody(unknownType, {})).to.throw(VError, `Unable to create body for unknown type: ${unknownType}`);
     });
   });
 });